Structural improvement by splitting JavaScript into separate module files
