+++ web_static/index.html
@@ -36,13 +36,22 @@
                 <div class="profile-details">
                     <!-- User profile details -->
                 </div>
+                <div class="change-password-form">
+                    <!-- Change password form elements -->
+                </div>
+                <div class="delete-account">
+                    <!-- Delete account section -->
+                </div>
             </div>
         </main>
     </div>
     <footer>
         <!-- Add footer content here -->
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="./scripts/dashboard.js"></script>
+    <script type="module" src="./scripts/expenses.js"></script>
+    <script type="module" src="./scripts/profile.js"></script>
+    <script type="module" src="./scripts/app.js"></script>
 </body>
 
 </html>
