+++ web_static/scripts/script.js
